requirements.txt
customtkinter
pydantic-ai
openai
google-generativeai
python-dotenv
